+++ Scripts/Spells/Spellweaving/ArcanistSpell.cs
@@ -13,7 +13,7 @@ namespace Server.Spells.Spellweaving
     {
     }
 
-    public abstract double RequiredSkill{ get; }
+    public abstract int RequiredSkill{ get; }
     public abstract int RequiredMana{ get; }
 
     public override SkillName CastSkill => SkillName.Spellweaving;
